Switching to using AsyncLocal<> instead of ThreadStatic
